refactor(storage)!: introduce `builder` module
