+++ src/integration-tests/tests/driver.rs
@@ -117,7 +117,7 @@ mod driver {
     #[test_case(Storage::builder().with_tracing().with_retry_policy(retry_policy()); "with tracing and retry enabled")]
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     async fn run_storage_objects(
-        builder: storage::client::ClientBuilder,
+        builder: storage::builder::storage::ClientBuilder,
     ) -> integration_tests::Result<()> {
         integration_tests::storage::objects(builder)
             .await
@@ -127,7 +127,7 @@ mod driver {
     #[test_case(Storage::builder().with_tracing().with_retry_policy(retry_policy()); "with tracing and retry enabled")]
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     async fn run_storage_objects_large_file(
-        builder: storage::client::ClientBuilder,
+        builder: storage::builder::storage::ClientBuilder,
     ) -> integration_tests::Result<()> {
         integration_tests::storage::objects_large_file(builder)
             .await
@@ -137,7 +137,7 @@ mod driver {
     #[test_case(Storage::builder().with_tracing().with_retry_policy(retry_policy()); "with tracing and retry enabled")]
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     async fn run_storage_objects_upload_buffered(
-        builder: storage::client::ClientBuilder,
+        builder: storage::builder::storage::ClientBuilder,
     ) -> integration_tests::Result<()> {
         integration_tests::storage::objects_upload_buffered(builder)
             .await
@@ -147,7 +147,7 @@ mod driver {
     #[test_case(Storage::builder().with_tracing().with_retry_policy(retry_policy()); "with tracing and retry enabled")]
     #[tokio::test(flavor = "multi_thread", worker_threads = 1)]
     async fn run_storage_objects_with_key(
-        builder: storage::client::ClientBuilder,
+        builder: storage::builder::storage::ClientBuilder,
     ) -> integration_tests::Result<()> {
         integration_tests::storage::objects_customer_supplied_encryption(builder)
             .await
