+++ src/storage/src/storage/upload_object/unbuffered.rs
@@ -94,9 +94,8 @@ where
 
 #[cfg(test)]
 mod tests {
+    use super::super::client::tests::{create_key_helper, test_inner_client};
     use super::*;
-    use crate::client::tests::create_key_helper;
-    use crate::client::tests::test_inner_client;
     use crate::upload_source::tests::VecStream;
     use http_body_util::BodyExt;
     use httptest::{Expectation, Server, matchers::*, responders::status_code};
